guest removal confirmation

+++ app/src/main/java/com/lais/study/ui/guests/ViewHolder.kt
@@ -1,23 +1,33 @@
 package com.lais.study.ui.guests
 
-import android.view.View
-import android.widget.TextView
-import androidx.appcompat.view.menu.MenuView.ItemView
+import android.app.AlertDialog
+import android.app.Dialog
 import androidx.recyclerview.widget.RecyclerView
 import com.lais.study.GuestModel
-import com.lais.study.R
 import com.lais.study.databinding.RowGuestsBinding
 
-class ViewHolder(private val bind: RowGuestsBinding, private val listener: OnGuestListener ): RecyclerView.ViewHolder(bind.root) {
+class ViewHolder(private val bind: RowGuestsBinding, private val listener: OnGuestListener) :
+    RecyclerView.ViewHolder(bind.root) {
 
-    fun bind(guest: GuestModel){
+    fun bind(guest: GuestModel) {
         bind.nameGuest.text = guest.name
-        bind.nameGuest.setOnClickListener{
+        bind.nameGuest.setOnClickListener {
             listener.onClick(guest.id)
         }
 
-        bind.nameGuest.setOnLongClickListener{
-            listener.onDelete(guest.id)
+        bind.nameGuest.setOnLongClickListener {
+
+            AlertDialog.Builder(itemView.context)
+                .setTitle("Remoção de convidado")
+                .setMessage("Tem certeza que deseja remover?")
+                .setPositiveButton(
+                    "Sim") { dialog, which ->
+                    listener.onDelete(guest.id)
+                }
+                .setNegativeButton("Não", null)
+                .create()
+                .show()
+
             true
         }
     }
